adds changes in some files
