+++ src/styles/Filters.css
@@ -8,6 +8,7 @@
   
   .button-row {
     display: flex;
+    flex-direction: column;
     justify-content: center;
     gap: 10px;
   }
@@ -16,6 +17,7 @@
     max-width: 100%;
     width: auto;
     display: flex;
+    flex-direction: column;
     justify-content: center;
     gap: 10px;
   }
@@ -59,6 +61,7 @@
   
   .block-average{
     display: flex;
+    flex-direction: column;
     justify-content: center; /* Espacio entre los contenedores */
     align-items: center; /* Alineación central de los contenedores */
     gap: 20px; /* Espacio entre los contenedores */
